add some basic benchmarks
