+++ wasm/testdata/hello-swift/Package.swift
@@ -1,0 +1,18 @@
+// swift-tools-version:5.3
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "hello-swift",
+    dependencies: [
+        .package(name: "Suborbital", path: "../../../")
+    ],
+    targets: [
+        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
+        // Targets can depend on other targets in this package, and on products in packages this package depends on.
+        .target(
+            name: "hello-swift",
+            dependencies: ["Suborbital"])
+    ]
+)
