+++ wasm/testdata/hello-swift/Sources/hello-swift/main.swift
@@ -1,0 +1,12 @@
+import Suborbital
+
+class HelloSwift: Suborbital.Runnable {
+    func run(input: String) -> String {
+        return "hello " + input
+    }
+}
+
+@_cdecl("init")
+func `init`() {
+    Suborbital.Set(runnable: HelloSwift())
+}
