--- wasm/testdata/swift-echo/.gitignore
@@ -1,5 +1,0 @@
-.DS_Store
-/.build
-/Packages
-/*.xcodeproj
-xcuserdata/
