--- wasm/testdata/swift-echo/.runnable.yaml
@@ -1,4 +1,0 @@
-name: swift-echo
-namespace: default
-lang: swift
-apiVersion: 0.2.5
