+++ wasm/testdata/swift-log/.runnable.yaml
@@ -1,0 +1,4 @@
+name: swift-log
+namespace: default
+lang: swift
+apiVersion: 0.2.5
